chore: replace AdaptiveLayout with SplitLayout
