--- composeApp/src/commonMain/kotlin/fr/nicopico/petitboutiste/ui/components/foundation/DragHandle.kt
@@ -1,58 +1,0 @@
-package fr.nicopico.petitboutiste.ui.components.foundation
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.DragHandle
-import androidx.compose.material3.Icon
-import androidx.compose.material3.LocalMinimumInteractiveComponentSize
-import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
-import androidx.compose.material3.adaptive.layout.PaneExpansionState
-import androidx.compose.material3.adaptive.layout.ThreePaneScaffoldScope
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-@Composable
-fun ThreePaneScaffoldScope.DragHandle(
-    state: PaneExpansionState,
-    contentDescription: String,
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    Box(
-        modifier =
-            Modifier
-                .paneExpansionDraggable(
-                    state,
-                    LocalMinimumInteractiveComponentSize.current,
-                    interactionSource,
-                    semanticsProperties = {},
-                )
-                .fillMaxHeight(),
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxHeight()
-                .width(1.dp)
-                .padding(vertical = 8.dp)
-                .background(Color.LightGray)
-                .align(Alignment.Center)
-        )
-        Icon(
-            Icons.Filled.DragHandle,
-            contentDescription,
-            modifier = Modifier
-                .rotate(90f)
-                .align(Alignment.Center)
-        )
-    }
-}
